In vehicle routing notebook, make Location + Depot problem facts and include them in solution
With the jpyinterpreter work done, Location and Depot need to be known
to OptaPy so cloning is done correctly. In particular, because
Location is used in a map, and Location does not define
__eq__ or __hash__, a Location can only be translated to its Java
counterpart once (otherwise, the location cannot be found in the
map, since the one in the map would be a different instance
than the location queried).